SID 0x86 formulas moved
