+++ uds/translator/data_record_definitions/__init__.py
@@ -39,8 +39,12 @@ from .conditional import (
     EVENT_TYPE_RECORD_03_2020,
     EVENT_TYPE_RECORD_07_2013,
     EVENT_TYPE_RECORD_07_2020,
+    EVENT_TYPE_RECORD_08_2020,
     EVENT_TYPE_RECORD_09_2020,
+    EVENT_WINDOW_TIME_2013,
+    EVENT_WINDOW_TIME_2020,
     SCALING_DATA_RECORDS,
+    SERVICE_TO_RESPOND,
 )
 from .did import (
     DATA_FROM_DID_2013,
@@ -116,9 +120,6 @@ from .other import (
     EPHEMERAL_PUBLIC_KEY_CLIENT_LENGTH,
     EPHEMERAL_PUBLIC_KEY_SERVER_LENGTH,
     EVENT_TYPE_RECORD_02,
-    EVENT_TYPE_RECORD_08_2020,
-    EVENT_WINDOW_TIME_2013,
-    EVENT_WINDOW_TIME_2020,
     FILE_AND_PATH_NAME_LENGTH,
     FILE_POSITION,
     FILE_SIZE_OR_DIR_INFO_PARAMETER_LENGTH,
@@ -142,7 +143,6 @@ from .other import (
     PROOF_OF_OWNERSHIP_SERVER_LENGTH,
     SECURITY_DATA_REQUEST_RECORD,
     SECURITY_DATA_RESPONSE_RECORD,
-    SERVICE_TO_RESPOND,
     SESSION_KEY_INFO_LENGTH,
     SESSION_PARAMETER_RECORD,
     SIGNATURE_ENCRYPTION_CALCULATION,
