+++ uds/translator/data_record_definitions/other.py
@@ -59,10 +59,7 @@ __all__ = [
     "NUMBER_OF_IDENTIFIED_EVENTS", "NUMBER_OF_ACTIVATED_EVENTS",
     "COMPARISON_LOGIC", "COMPARE_VALUE", "HYSTERESIS_VALUE",
     "COMPARE_SIGN", "BITS_NUMBER", "BIT_OFFSET", "LOCALIZATION",
-    "EVENT_WINDOW_TIME_2013", "EVENT_WINDOW_TIME_2020",
-    "EVENT_TYPE_RECORD_08_2020",
     "EVENT_TYPE_RECORD_02",
-    "SERVICE_TO_RESPOND",
     # SID 0x87
     "LINK_RECORD",
     "LINK_CONTROL_MODE_IDENTIFIER",
@@ -70,7 +67,7 @@ __all__ = [
 ]
 
 
-from typing import Optional, Tuple, Union
+from typing import Optional
 
 from uds.utilities import (
     AUTHENTICATION_RETURN_PARAMETER_MAPPING,
@@ -106,7 +103,6 @@ from uds.utilities import (
 )
 
 from ..data_record import (
-    ConditionalFormulaDataRecord,
     ConditionalMappingDataRecord,
     CustomFormulaDataRecord,
     LinearFormulaDataRecord,
@@ -114,121 +110,9 @@ from ..data_record import (
     MappingDataRecord,
     RawDataRecord,
 )
-from .sub_functions import EVENT_TYPE_2013, EVENT_TYPE_2020, REPORT_TYPE_2020
-
-# Formulas
-
-
-
-
-
-
-
-
-
-
-def get_event_window_2013(event_number: Optional[int] = None) -> MappingDataRecord:
-    """
-    Get eventWindowTime Data Record compatible with ISO 14229-1:2013 version.
-
-    :param event_number: Order number of the event record to contain this Data Record.
-        None if there are no records.
-
-    :return: Created eventWindowTime Data Record.
-    """
-    return MappingDataRecord(name="eventWindowTime" if event_number is None else f"eventWindowTime#{event_number}",
-                             length=8,
-                             values_mapping=EVENT_WINDOW_TIME_MAPPING_2013)
-
-
-def get_event_window_2020(event_number: Optional[int] = None) -> MappingDataRecord:
-    """
-    Get eventWindowTime Data Record compatible with ISO 14229-1:2020 version.
-
-    :param event_number: Order number of the event record to contain this Data Record.
-        None if there are no records.
-
-    :return: Created eventWindowTime Data Record.
-    """
-    return MappingDataRecord(name="eventWindowTime" if event_number is None else f"eventWindowTime#{event_number}",
-                             length=8,
-                             values_mapping=EVENT_WINDOW_TIME_MAPPING_2020)
-
-
-def get_service_to_respond(event_number: Optional[int] = None) -> RawDataRecord:
-    """
-    Get serviceToRespondToRecord Data Record.
-
-    :param event_number: Order number of the event record to contain this Data Record.
-        None if there are no records.
-
-    :return: Created serviceToRespondToRecord Data Record.
-    """
-    return RawDataRecord(name="serviceToRespondToRecord" if event_number is None
-                              else f"serviceToRespondToRecord#{event_number}",
-                         length=8,
-                         min_occurrences=1,
-                         max_occurrences=None)
-
-
-def get_event_type_of_active_event_2013(event_number: int) -> RawDataRecord:
-    """
-    Get eventTypeOfActiveEvent Data Record.
-
-    :param event_number: Number of the active event.
-
-    :return: Created eventTypeOfActiveEvent Data Record.
-    """
-    return RawDataRecord(name=f"eventTypeOfActiveEvent#{event_number}",
-                         length=8,
-                         children=(RESERVED_BIT,
-                                   EVENT_TYPE_2013))
-
-
-def get_event_type_of_active_event_2020(event_number: int) -> RawDataRecord:
-    """
-    Get eventTypeOfActiveEvent Data Record.
-
-    :param event_number: Number of the active event.
-
-    :return: Created eventTypeOfActiveEvent Data Record.
-    """
-    return RawDataRecord(name=f"eventTypeOfActiveEvent#{event_number}",
-                         length=8,
-                         children=(RESERVED_BIT,
-                                   EVENT_TYPE_2020))
-
-
-def get_event_type_record_02(event_number: Optional[int] = None) -> RawDataRecord:
-    """
-    Get eventTypeRecord Data Record for event equal to 0x02.
-
-    :param event_number: Order number of the event record to contain this Data Record.
-        None if there are no records.
-
-    :return: Created eventTypeRecord Data Record.
-    """
-    return RawDataRecord(name="eventTypeRecord" if event_number is None else f"eventTypeRecord#{event_number}",
-                         length=8,
-                         children=(TIMER_SCHEDULE,))
-
-
-def get_event_type_record_08_2020(event_number: Optional[int] = None) -> RawDataRecord:
-    """
-    Get eventTypeRecord Data Record for event equal to 0x08.
-
-    :param event_number: Order number of the event record to contain this Data Record.
-        None if there are no records.
-
-    :return: Created eventTypeRecord Data Record.
-    """
-    return RawDataRecord(name="eventTypeRecord" if event_number is None else f"eventTypeRecord#{event_number}",
-                         length=8,
-                         children=(RESERVED_BIT,
-                                   REPORT_TYPE_2020))
-
 
 # Shared
+
 RESERVED_BIT = RawDataRecord(name="reserved",
                              length=1)
 RESERVED_2BITS = RawDataRecord(name="reserved",
@@ -628,20 +512,16 @@ LOCALIZATION = RawDataRecord(name="Localization",
                              children=(COMPARE_SIGN,
                                        BITS_NUMBER,
                                        BIT_OFFSET))
-EVENT_WINDOW_TIME_2020 = get_event_window_2020()
-EVENT_WINDOW_TIME_2013 = get_event_window_2013()
 
-EVENT_TYPE_RECORD_08_2020 = get_event_type_record_08_2020()
 
-TIMER_SCHEDULE = MappingDataRecord(name="Timer schedule",
-                                   length=8,
-                                   values_mapping=TIMER_SCHEDULE_MAPPING_2013)
+TIMER_SCHEDULE_2013 = MappingDataRecord(name="Timer schedule",
+                                        length=8,
+                                        values_mapping=TIMER_SCHEDULE_MAPPING_2013)
 
 EVENT_TYPE_RECORD_02 = RawDataRecord(name="eventTypeRecord",
                                      length=8,
-                                     children=(TIMER_SCHEDULE,))
+                                     children=(TIMER_SCHEDULE_2013,))
 
-SERVICE_TO_RESPOND = get_service_to_respond()
 
 # SID 0x87
 LINK_CONTROL_MODE_IDENTIFIER = MappingDataRecord(name="linkControlModeIdentifier",
